Bugfix for concrete value

+++ src/edu/sharif/twinner/trace/ConcreteValueAbstractImp.h
@@ -60,7 +60,7 @@ public:
 
   friend std::basic_ostream<char> &operator<< (std::basic_ostream<char> &stream,
       const ConcreteValueAbstractImp<bits, ValueType> &me) {
-    return stream << "0x" << std::hex << me.value;
+    return stream << "0x" << std::hex << (unsigned int) me.value;
   }
 
   virtual bool isNegative () const {
